Graphcore/benchmarks/bert/implementations/paddlepaddle/requirements.txt
# paddle
paddlenlp==2.2.6
git+https://github.com/mlperf/logging.git@2.0.0-rc3
numpy
numba
mpi4py
tqdm
wandb
tensorflow==1.15.5